Move validateDFUSupport to a more central module
