+++ src/cmd/update.js
@@ -2,12 +2,12 @@ const ParticleApi = require('./api');
 const { platformForId } = require('../lib/platform');
 const settings = require('../../settings');
 const semver = require('semver');
-
 const usbUtils = require('./usb-util');
 const deviceOsUtils = require('../lib/device-os-version-util');
 const CLICommandBase = require('./base');
-const { parseModulesToFlash, filterModulesToFlash, maintainDeviceProtection, createFlashSteps, flashFiles, validateDFUSupport } = require('../lib/flash-helper');
+const { parseModulesToFlash, filterModulesToFlash, maintainDeviceProtection, createFlashSteps, flashFiles } = require('../lib/flash-helper');
 const createApiCache = require('../lib/api-cache');
+const { validateDFUSupport } = require('./device-util');
 
 module.exports = class UpdateCommand extends CLICommandBase {
