+++ src/cmd/usb-util.js
@@ -12,7 +12,7 @@ const {
 	DeviceProtectionError
 } = require('particle-usb');
 const deviceProtectionHelper = require('../lib/device-protection-helper');
-const { validateDFUSupport } = require('../lib/flash-helper');
+const { validateDFUSupport } = require('./device-util');
 
 // Timeout when reopening a USB device after an update via control requests. This timeout should be
 // long enough to allow the bootloader apply the update
